Move nonce out of core into own package.
