+++ wfe/web-front-end_test.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/cactus/go-statsd-client/statsd"
 	"github.com/jmhodges/clock"
+	"github.com/letsencrypt/boulder/nonce"
 	"github.com/letsencrypt/boulder/probs"
 	"github.com/square/go-jose"
 
@@ -188,7 +189,7 @@ func makeBody(s string) io.ReadCloser {
 	return ioutil.NopCloser(strings.NewReader(s))
 }
 
-func signRequest(t *testing.T, req string, nonceService *core.NonceService) string {
+func signRequest(t *testing.T, req string, nonceService *nonce.NonceService) string {
 	accountKey, err := jose.LoadPrivateKey([]byte(test1KeyPrivatePEM))
 	test.AssertNotError(t, err, "Failed to load key")
